refactor metadata as module
