+++ astropy/utils/metadata/exceptions.py
@@ -1,0 +1,15 @@
+# Licensed under a 3-clause BSD style license - see LICENSE.rst
+"""Metadata exceptions and warnings."""
+
+
+from astropy.utils.exceptions import AstropyWarning
+
+__all__ = []
+
+
+class MergeConflictError(TypeError):
+    pass
+
+
+class MergeConflictWarning(AstropyWarning):
+    pass
