+++ docs/changes/utils/15166.api.rst
@@ -1,0 +1,2 @@
+The structure of ``utils.meta`` has been refactored, but all the available
+functions and classes are still present and should be imported as before.
